Update gryke.py
Included the import of the curve_fit functions.
Included the profile_building function, which is needed by gryke.
Included the dike_mode_section function, which is used to plot the geometrically modelled dike with gryke underneath the topography.
gryke core function updated so that it includes the identifier as the first argument. This identifier is needed by the dike_model_section function.
Improved the descriptions of the functions.